Changed RabbitMqQueueCheck to RabbitMqManagementQueueCheck
