+++ docker-compose.yml
@@ -11,7 +11,7 @@ services:
             REDIS_HOST: 'diagnostic-redis'
             MAILER_HOST: 'diagnostic-mailcatcher'
             ELASTICSEARCH_HOST: 'elasticsearch'
-            AWS_DYNAMODB_ENDPOINT: 'http://diagnostic-dynamodb:8000'
+            AWS_DYNAMODB_ENDPOINT: 'http://dynamodb:8000'
         volumes:
             - './:/code'
         networks:
