+++ src/Check/RabbitMq/AmqpExt/RabbitMqConnectionCheck.php
@@ -13,25 +13,69 @@ declare(strict_types = 1);
 
 namespace FiveLab\Component\Diagnostic\Check\RabbitMq\AmqpExt;
 
+use FiveLab\Component\Diagnostic\Check\CheckInterface;
+use FiveLab\Component\Diagnostic\Check\RabbitMq\RabbitMqConnectionParameters;
+use FiveLab\Component\Diagnostic\Result\Failure;
 use FiveLab\Component\Diagnostic\Result\ResultInterface;
 use FiveLab\Component\Diagnostic\Result\Success;
 
 /**
  * Check connect to RabbitMQ
  */
-class RabbitMqConnectionCheck extends AbstractRabbitMqCheck
+class RabbitMqConnectionCheck implements CheckInterface
 {
+    /**
+     * @var RabbitMqConnectionParameters
+     */
+    private $connectionParameters;
+
+    /**
+     * Constructor.
+     *
+     * @param RabbitMqConnectionParameters $connectionParameters
+     */
+    public function __construct(RabbitMqConnectionParameters $connectionParameters)
+    {
+        $this->connectionParameters = $connectionParameters;
+    }
+
     /**
      * {@inheritdoc}
      */
     public function check(): ResultInterface
     {
-        $failure = $this->connect();
+        if (!\class_exists(\AMQPConnection::class)) {
+            return new Failure('The ext-amqp not installed.');
+        }
 
-        if ($failure) {
-            return $failure;
+        $connectionParameters = [
+            'host'            => $this->connectionParameters->getHost(),
+            'port'            => $this->connectionParameters->getPort(),
+            'vhost'           => $this->connectionParameters->getVhost(),
+            'login'           => $this->connectionParameters->getUsername(),
+            'password'        => $this->connectionParameters->getPassword(),
+            'connect_timeout' => 5,
+        ];
+
+        $connection = new \AMQPConnection($connectionParameters);
+
+        try {
+            $connection->connect();
+        } catch (\AMQPConnectionException $e) {
+            return new Failure($e->getMessage());
         }
 
         return new Success('Success connect to RabbitMQ');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getExtraParameters(): array
+    {
+        return [
+            'dsn'   => $this->connectionParameters->getDsn(false, true),
+            'vhost' => $this->connectionParameters->getVhost(),
+        ];
     }
 }
