+++ src/Check/RabbitMq/Management/RabbitMqManagementQueueCheck.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace FiveLab\Component\Diagnostic\Check\RabbitMq\Management;
 
@@ -18,11 +18,13 @@ use FiveLab\Component\Diagnostic\Check\RabbitMq\RabbitMqConnectionParameters;
 use FiveLab\Component\Diagnostic\Result\Failure;
 use FiveLab\Component\Diagnostic\Result\ResultInterface;
 use FiveLab\Component\Diagnostic\Result\Success;
+use FiveLab\Component\Diagnostic\Result\Warning;
 use Http\Client\HttpClient;
 use Http\Discovery\HttpClientDiscovery;
 use Http\Discovery\Psr17FactoryDiscovery;
 use Http\Message\RequestFactory;
 use Psr\Http\Client\ClientExceptionInterface;
+use Psr\Http\Message\ResponseInterface;
 
 /**
  * Check queue existence
@@ -50,17 +52,43 @@ class RabbitMqManagementQueueCheck implements CheckInterface
     private $queueName;
 
     /**
-     * Constructor.
-     *
+     * @var int|null
+     */
+    private $maxMessages;
+
+    /**
+     * must be between 0 and 100
+     * @var int
+     */
+    private $maxWarningPercentage;
+
+    /**
+     * @var int|null
+     */
+    private $minMessages;
+
+    /**
      * @param RabbitMqConnectionParameters $connectionParameters
      * @param string                       $queueName
+     * @param int|null                     $maxMessages
+     * @param int|null                     $minMessages
+     * @param int|null                     $maxWarningPercentage
      * @param HttpClient|null              $client
      * @param RequestFactory|null          $requestFactory
      */
-    public function __construct(RabbitMqConnectionParameters $connectionParameters, string $queueName, HttpClient $client = null, RequestFactory $requestFactory = null)
+    public function __construct(RabbitMqConnectionParameters $connectionParameters, string $queueName, int $maxMessages = null, int $minMessages = null, int $maxWarningPercentage = null, HttpClient $client = null, RequestFactory $requestFactory = null)
     {
         $this->connectionParameters = $connectionParameters;
         $this->queueName = $queueName;
+        $this->maxMessages = $maxMessages;
+        $this->minMessages = $minMessages;
+
+        if ($maxWarningPercentage < 0 || $maxWarningPercentage > 100) {
+            throw new \InvalidArgumentException('$maxWarningPercentage must be between 0 and 100');
+        }
+
+        $this->maxWarningPercentage = $maxWarningPercentage;
+
         $this->client = $client ?: HttpClientDiscovery::find();
         $this->requestFactory = $requestFactory ?: Psr17FactoryDiscovery::findRequestFactory();
     }
@@ -101,6 +129,11 @@ class RabbitMqManagementQueueCheck implements CheckInterface
             ));
         }
 
+        $result = $this->checkMessageLimits($response);
+        if ($result) {
+            return $result;
+        }
+
         return new Success('Success check queue via RabbitMQ Management API.');
     }
 
@@ -110,9 +143,91 @@ class RabbitMqManagementQueueCheck implements CheckInterface
     public function getExtraParameters(): array
     {
         return [
-            'dsn'   => $this->connectionParameters->getDsn(true, true),
-            'vhost' => $this->connectionParameters->getVhost(),
-            'queue' => $this->queueName,
+            'dsn'                    => $this->connectionParameters->getDsn(true, true),
+            'vhost'                  => $this->connectionParameters->getVhost(),
+            'queue'                  => $this->queueName,
+            'max_messages'           => $this->maxMessages,
+            'min_messages'           => $this->minMessages,
+            'max_warning_percentage' => $this->maxWarningPercentage,
         ];
+    }
+
+    /**
+     * @param ResponseInterface $response
+     *
+     * @return ResultInterface|null
+     */
+    private function checkMessageLimits(ResponseInterface $response): ?ResultInterface
+    {
+        if (!\is_int($this->maxMessages) && !\is_int($this->minMessages)) {
+            return null;
+        }
+
+        $queueDetails = \json_decode((string) $response->getBody(), true);
+
+        $queuedMessages = $queueDetails['messages'] ?? 0;
+
+        $maxMessagesResult = $this->checkForMaxMessages($queuedMessages);
+        if ($maxMessagesResult) {
+            return $maxMessagesResult;
+        }
+
+        $minMessagesResult = $this->checkForMinMessages($queuedMessages);
+        if ($minMessagesResult) {
+            return $minMessagesResult;
+        }
+
+        return null;
+    }
+
+    /**
+     * @param int $queuedMessages
+     *
+     * @return ResultInterface|null
+     */
+    private function checkForMaxMessages(int $queuedMessages): ?ResultInterface
+    {
+        if (!\is_int($this->maxMessages)) {
+            return null;
+        }
+
+        switch (true) {
+            case $queuedMessages > $this->maxMessages:
+                return new Failure(\sprintf(
+                    '%d messages found! Max allowed %d for queue %s',
+                    $queuedMessages,
+                    $this->maxMessages,
+                    $this->queueName
+                ));
+
+            case $this->maxWarningPercentage && ($queuedMessages > ($this->maxMessages * $this->maxWarningPercentage) / 100):
+                return new Warning(\sprintf(
+                    'Warning! %d messages found. Max %d for queue %s',
+                    $queuedMessages,
+                    $this->maxMessages,
+                    $this->queueName
+                ));
+        }
+
+        return null;
+    }
+
+    /**
+     * @param int $queuedMessages
+     *
+     * @return ResultInterface|null
+     */
+    private function checkForMinMessages(int $queuedMessages): ?ResultInterface
+    {
+        if (!\is_int($this->minMessages) || $queuedMessages >= $this->minMessages) {
+            return null;
+        }
+
+        return new Failure(\sprintf(
+            '%d messages found! Minimum required %d for queue %s',
+            $queuedMessages,
+            $this->minMessages,
+            $this->queueName
+        ));
     }
 }
